+++ plugins/module_utils/_auth_method_none.py
@@ -1,0 +1,31 @@
+# -*- coding: utf-8 -*-
+# Copyright (c) 2021 Brian Scholer (@briantist)
+# Simplified BSD License (see licenses/simplified_bsd.txt or https://opensource.org/licenses/BSD-2-Clause)
+
+'''Python versions supported: all controller-side versions, all remote-side versions except 2.6'''
+
+# FOR INTERNAL COLLECTION USE ONLY
+# The interfaces in this file are meant for use within the community.hashi_vault collection
+# and may not remain stable to outside uses. Changes may be made in ANY release, even a bugfix release.
+# See also: https://github.com/ansible/community/issues/539#issuecomment-780839686
+# Please open an issue if you have questions about this.
+
+from __future__ import absolute_import, division, print_function
+__metaclass__ = type
+
+from ansible_collections.community.hashi_vault.plugins.module_utils._hashi_vault_common import HashiVaultAuthMethodBase
+
+class HashiVaultAuthMethodNone(HashiVaultAuthMethodBase):
+    '''HashiVault option group class for auth: none'''
+
+    NAME = 'none'
+    OPTIONS = []
+
+    def __init__(self, option_adapter, warning_callback):
+        super(HashiVaultAuthMethodNone, self).__init__(option_adapter, warning_callback)
+
+    def validate(self):
+        pass
+
+    def authenticate(self, client, use_token=False):
+        return None
